Add files via upload
or this, we are going to take the data from Sunnybrook cardiac data(SCD), as this comes with
expertly labeled ground-truth LV contours. This dataset is available in the cloud atlas project website,
basically curated from The National Heart, Lung, and Blood Institute (NHLBI).
Here the MRI is saved in .dcm format or DICOM((Digital Imaging and Communications in
Medicine) format. There are three different types of views that an MRI Image can take in order to
examine the heart. These views are the 4-Chamber view, 2-Chamber view, and the Short Axis view.
Here the focus is on using the Short Axis (SAX) views. For each patient, we have a number of
30-frame MRIs in the ​ DICOM​ format, showing the heart during a single cardiac cycle.